requirements.txt
apache-airflow==2.9.0

pandas
numpy

sqlalchemy
psycopg2-binary

python-dotenv
